vacuum entity cleanup
